+++ Worksheet Features/Fill Series/.NET/DateTimeFillSeries/DateTimeFillSeries.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.14.36127.28 d17.14
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DateTime FillSeries", "DateTime FillSeries\DateTime FillSeries.csproj", "{EA9A1164-D806-4836-AC67-CC2AF5816704}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DateTimeFillSeries", "DateTimeFillSeries\DateTimeFillSeries.csproj", "{346AE701-4612-43C5-AD9A-3E2789A1CD32}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Syncfusion.XlsIO.Portable_NET80", "..\..\..\..\..\xlsio\Portable\XlsIO.Portable\Src\Syncfusion.XlsIO.Portable_NET80.csproj", "{B1829367-CB57-24F6-0C45-DBCA10D321FF}"
 EndProject
@@ -15,10 +15,10 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{EA9A1164-D806-4836-AC67-CC2AF5816704}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{EA9A1164-D806-4836-AC67-CC2AF5816704}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{EA9A1164-D806-4836-AC67-CC2AF5816704}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{EA9A1164-D806-4836-AC67-CC2AF5816704}.Release|Any CPU.Build.0 = Release|Any CPU
+		{346AE701-4612-43C5-AD9A-3E2789A1CD32}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{346AE701-4612-43C5-AD9A-3E2789A1CD32}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{346AE701-4612-43C5-AD9A-3E2789A1CD32}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{346AE701-4612-43C5-AD9A-3E2789A1CD32}.Release|Any CPU.Build.0 = Release|Any CPU
 		{B1829367-CB57-24F6-0C45-DBCA10D321FF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{B1829367-CB57-24F6-0C45-DBCA10D321FF}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{B1829367-CB57-24F6-0C45-DBCA10D321FF}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -32,6 +32,6 @@ Global
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {FAD9B7EE-3BB0-4977-8762-A415698A72BB}
+		SolutionGuid = {80FA51BF-09BC-42BE-8C5F-1A1D2108B0DE}
 	EndGlobalSection
 EndGlobal
