+++ Worksheet Features/Fill Series/.NET/DateTimeFillSeries/DateTimeFillSeries/DateTimeFillSeries.csproj
@@ -3,7 +3,6 @@
   <PropertyGroup>
     <OutputType>Exe</OutputType>
     <TargetFramework>net8.0</TargetFramework>
-    <RootNamespace>Number_FillSeries</RootNamespace>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
   </PropertyGroup>
