+++ Worksheet Features/Fill Series/.NET/DateTimeFillSeries/README.md
@@ -1,0 +1,42 @@
+# How to fill a date series in a worksheet?
+
+Step 1: Create a new C# Console Application project.
+
+Step 2: Name the project.
+
+Step 3: Install the [Syncfusion.XlsIO.Net.Core](https://www.nuget.org/packages/Syncfusion.XlsIO.Net.Core) NuGet package as reference to your .NET Standard applications from [NuGet.org](https://www.nuget.org).
+
+Step 4: Include the following namespaces in the **Program.cs** file.
+
+```csharp
+using System;
+using System.IO;
+using Syncfusion.XlsIO;
+```
+
+Step 5: Include the below code snippet in **Program.cs**  to fill a date series in a worksheet.
+```csharp
+using (ExcelEngine excelEngine = new ExcelEngine())
+{
+    IApplication application = excelEngine.Excel;
+    application.DefaultVersion = ExcelVersion.Xlsx;
+    IWorkbook workbook = application.Workbooks.Create(1);
+    IWorksheet worksheet = workbook.Worksheets[0];
+
+    //Assign datetime value to the cell
+    worksheet["A1"].DateTime = new DateTime(2025, 1, 1);
+
+    //Define the range
+    IRange range = worksheet["A1:A50"];
+
+    //Fill series using the years option
+    range.FillSeries(ExcelSeriesBy.Columns, ExcelFillSeries.Years, 2, new DateTime(2100, 1, 1));
+
+    //Saving the workbook 
+    FileStream outputStream = new FileStream(Path.GetFullPath("Output/Output.xlsx"), FileMode.Create, FileAccess.Write);
+    workbook.SaveAs(outputStream);
+
+    //Dispose streams
+    outputStream.Dispose();
+}
+```
