+++ Worksheet Features/Fill Series/.NET/FillSeriesByEnablingTrend/FillSeriesByEnablingTrend.sln
@@ -3,11 +3,11 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.14.36127.28 d17.14
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FillSeries by Enabling Trend", "FillSeries by Enabling Trend\FillSeries by Enabling Trend.csproj", "{B94BB8DF-E6D3-4F7B-B908-E8DDB53E9FF2}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Syncfusion.XlsIO.Portable_NET80", "..\..\..\..\..\xlsio\Portable\XlsIO.Portable\Src\Syncfusion.XlsIO.Portable_NET80.csproj", "{B1829367-CB57-24F6-0C45-DBCA10D321FF}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FillSeriesByEnablingTrend", "FillSeriesByEnablingTrend\FillSeriesByEnablingTrend.csproj", "{9514937A-D01E-463C-99EE-58739CF69460}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Syncfusion.Compression.Portable_NET80", "..\..\..\..\..\compression\Portable\Compression.Portable\Src\Syncfusion.Compression.Portable_NET80.csproj", "{652B9342-F913-C1BF-A30E-31072C9225F7}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Syncfusion.XlsIO.Portable_NET80", "..\..\..\..\..\xlsio\Portable\XlsIO.Portable\Src\Syncfusion.XlsIO.Portable_NET80.csproj", "{B1829367-CB57-24F6-0C45-DBCA10D321FF}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,23 +15,23 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{B94BB8DF-E6D3-4F7B-B908-E8DDB53E9FF2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{B94BB8DF-E6D3-4F7B-B908-E8DDB53E9FF2}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{B94BB8DF-E6D3-4F7B-B908-E8DDB53E9FF2}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{B94BB8DF-E6D3-4F7B-B908-E8DDB53E9FF2}.Release|Any CPU.Build.0 = Release|Any CPU
-		{B1829367-CB57-24F6-0C45-DBCA10D321FF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{B1829367-CB57-24F6-0C45-DBCA10D321FF}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{B1829367-CB57-24F6-0C45-DBCA10D321FF}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{B1829367-CB57-24F6-0C45-DBCA10D321FF}.Release|Any CPU.Build.0 = Release|Any CPU
+		{9514937A-D01E-463C-99EE-58739CF69460}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9514937A-D01E-463C-99EE-58739CF69460}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9514937A-D01E-463C-99EE-58739CF69460}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9514937A-D01E-463C-99EE-58739CF69460}.Release|Any CPU.Build.0 = Release|Any CPU
 		{652B9342-F913-C1BF-A30E-31072C9225F7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{652B9342-F913-C1BF-A30E-31072C9225F7}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{652B9342-F913-C1BF-A30E-31072C9225F7}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{652B9342-F913-C1BF-A30E-31072C9225F7}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B1829367-CB57-24F6-0C45-DBCA10D321FF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B1829367-CB57-24F6-0C45-DBCA10D321FF}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B1829367-CB57-24F6-0C45-DBCA10D321FF}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B1829367-CB57-24F6-0C45-DBCA10D321FF}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {45D78520-D413-42CC-B68D-09C8135BDEE5}
+		SolutionGuid = {3292C768-1ED4-41AF-B4D0-966FE00D1E76}
 	EndGlobalSection
 EndGlobal
