+++ Worksheet Features/Fill Series/.NET/NumberFillSeries/NumberFillSeries/NumberFillSeries.csproj
@@ -3,7 +3,6 @@
   <PropertyGroup>
     <OutputType>Exe</OutputType>
     <TargetFramework>net8.0</TargetFramework>
-    <RootNamespace>FillSeries_by_Enabling_Trend</RootNamespace>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
   </PropertyGroup>
