+++ Worksheet Features/Fill Series/.NET/NumberFillSeries/README.md
@@ -1,4 +1,4 @@
-# How to fill series by enabling trend in a worksheet?
+# How to fill a number series in a worksheet?
 
 Step 1: Create a new C# Console Application project.
 
@@ -14,7 +14,7 @@ using System.IO;
 using Syncfusion.XlsIO;
 ```
 
-Step 5: Include the below code snippet in **Program.cs**  to fill series by enabling trend in a worksheet.
+Step 5: Include the below code snippet in **Program.cs**  to fill a number series in a worksheet.
 ```csharp
 using (ExcelEngine excelEngine = new ExcelEngine())
 {
@@ -23,19 +23,17 @@ using (ExcelEngine excelEngine = new ExcelEngine())
     IWorkbook workbook = application.Workbooks.Create(1);
     IWorksheet worksheet = workbook.Worksheets[0];
 
-    //Setting the values to the cells
+    //Assign value to the cell
     worksheet["A1"].Number = 1;
-    worksheet["A2"].Number = 3;
-    worksheet["A3"].Number = 2;
 
     //Define the range
-    IRange range = worksheet["A1:A10"];
+    IRange range = worksheet["A1:A100"];
 
-    //Use FillSeries method to fill the values by enabling trend
-    range.FillSeries(ExcelSeriesBy.Columns, ExcelFillSeries.Linear, true);
+    //Fill series using the linear option
+    range.FillSeries(ExcelSeriesBy.Columns, ExcelFillSeries.Linear, 5, 1000);
 
     //Saving the workbook 
-    FileStream outputStream = new FileStream(Path.GetFullPath("Output.xlsx"), FileMode.Create, FileAccess.Write);
+    FileStream outputStream = new FileStream(Path.GetFullPath("Output/Output.xlsx"), FileMode.Create, FileAccess.Write);
     workbook.SaveAs(outputStream);
 
     //Dispose streams
